% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biolockjVersion.R
\name{biolockjVersion}
\alias{biolockjVersion}
\title{BioLockJ Version}
\usage{
biolockjVersion(showBuild = FALSE)
}
\arguments{
\item{showBuild}{TRUE/FALSE. If TRUE, the version string will include the build hash string.}
}
\value{
String giving the current BioLockJ version
}
\description{
Get the version of BioLockJ. Optionally include the build string.
}
\examples{
biolockjVersion()
biolockjVersion(TRUE)

}
