% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callBioLockJ.R
\name{callBioLockJ}
\alias{callBioLockJ}
\title{Call the BioLockJ java program}
\usage{
callBioLockJ(args, externalModules = NULL)
}
\arguments{
\item{args}{args to pass to BioLockJ java program}

\item{externalModules}{file path to directory with external modules}
}
\value{
Any and all results that the BioLockJ java program prints to standard out
}
\description{
Call the BioLockJ java program
}
