% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callBioLockJApi.R
\name{callBioLockJApi}
\alias{callBioLockJApi}
\title{Call the helper API of BioLockJ java program}
\usage{
callBioLockJApi(args, external.modules = NULL)
}
\arguments{
\item{args}{args to pass to BioLockJ java program}

\item{external.modules}{file path to directory with external modules}
}
\value{
Any and all results that the BioLockJ java program prints to standard out
}
\description{
Call the helper API of BioLockJ java program
}
