% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_BioLockR.R, R/getBljJar.R,
%   R/getBljProj.R, R/setBljProj.R, R/set_BLJ_JAR.R
\docType{package}
\name{config_BioLockR}
\alias{config_BioLockR}
\alias{getBljJar}
\alias{getBljProj}
\alias{setBljProj}
\alias{set_BLJ_JAR}
\title{Configure BioLockR}
\usage{
getBljJar()

getBljProj()

setBljProj(path, remember = TRUE, doublecheck = TRUE)

set_BLJ_JAR(path, remember = TRUE, doublecheck = TRUE)
}
\arguments{
\item{path}{a file path to the BioLockJ jar file (set_BLJ_JAR()) to the BioLockJ project directory (setBljProj).}

\item{remember}{TRUE/FALSE: Should this value be remembered for future sessions.
If true, the value is appended to the $HOME/.Renviron file, which can be edited
using: usethis::edit_r_environ()}

\item{doublecheck}{(only used if remember==TRUE) should an interactive response be required before saving value to a file.}
}
\value{
"get" methods return a file path.
"set" methods, return boolean; TRUE if the value is set for that property

getBljJar: The path to the BioLockJ jar file

getBljProj: the path to the BioLockJ project directory

setBljProj: boolean; TRUE if path is set for current session

set_BLJ_JAR: boolean; TRUE if path is set for current session
}
\description{
Set BioLockJ Essential File Paths
<br>
set_BLJ_JAR and setBljProj set variables BLJ_JAR and BLJ_PROJ respectively.
These variables can also be set by adding them to your .Renviron file, which
can be edited using: usethis::edit_r_environ() .
}
\details{
If you have BioLockJ installed, you have a variable $BLJ, and likley $BLJ_RPOJ.
These may not be detected in an R session, or an Rstudio R session.
If these are detected, they are used; otherwise they must be set manually.
In a terminal session try: echo $BLJ
Copy this path and use it with set_BLJ_JAR().

In some cases, an R session launched from the terminal may have access to the
variables, while a session launched in Rstudio does not. In this case, open R from a terminal
and run:
library(BioLockR)
set_BLJ_JAR(getBljJar(), remember=TRUE)
Now subsequent sessions in Rstudio will have the jar path.

If you update your BioLockJ location, use set_BLJ_JAR("", remember=TRUE) and then repeat the above steps.
}
\section{Functions}{
\itemize{
\item \code{getBljJar}: get the BioLockJ.jar file path

\item \code{getBljProj}: get the BioLockJ projects directory path

\item \code{setBljProj}: set the BioLockJ projects directory path

\item \code{set_BLJ_JAR}: set the BioLockJ.jar file path
}}

\examples{
# set_BLJ_JAR("/path/to/BioLockJ/dist/BioLockJ.jar")
# getBljJar()

# setBljProj("~/Documents/BioLockJ/pipelines")
# getBljProj()
# >[1] "~/Documents/BioLockJ/pipelines"

setBljProj( "path/to/projects/dir" )

set_BLJ_JAR("path/to/BioLockJ/dist/BioLockJ.jar")

}
