% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config_file_path.R
\name{get_config_file_path}
\alias{get_config_file_path}
\title{Get BioLockR settings file}
\usage{
get_config_file_path()
}
\value{

}
\description{
Get the path to the file where the BioLockR package saves user preferences,
namely the path to BLJ_JAR and BLJ_PROJ.
}
