% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBljJar.R, R/getBljProj.R
\name{getBljJar}
\alias{getBljJar}
\alias{getBljProj}
\title{Get BioLockJ essential file paths}
\usage{
getBljJar()

getBljProj()
}
\value{
The path to the BioLockJ jar file

the path to the BioLockJ project directory
}
\description{
Get the essential file paths associated with the BioLockJ java program.
}
\details{
getBljJar returns the path to the jar file of the BioLockJ java program based on either BLJ_JAR (if set) or BLJ.

bljProj returns the path to the BioLockJ project directory (where new pipeline instances are created) based on environemnt variable BLJ_PROJ.
}
\section{Functions}{
\itemize{
\item \code{getBljProj}: 
}}

\examples{
getBljJar()

getBljProj()

}
\seealso{
setBljJar, setBljProj
}
