% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isReadableValue.R
\name{isReadableValue}
\alias{isReadableValue}
\title{Title}
\usage{
isReadableValue(value)
}
\arguments{
\item{value}{}
}
\value{

}
\description{
Title
}
\examples{
isReadableValue(NA)
#> FALSE
isReadableValue(c())
#> FALSE
isReadableValue("")
#> FALSE
isReadableValue("text")
#> TRUE

}
