% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_pipeline.R
\name{last_pipeline}
\alias{last_pipeline}
\title{Get most recent pipeline}
\usage{
last_pipeline()
}
\value{
top level file path of the most recently modified pipeline
}
\description{
Get most recent pipeline
}
