% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listApiModules.R
\name{listApiModules}
\alias{listApiModules}
\title{List Modules}
\usage{
listApiModules(externalModules = NULL)
}
\arguments{
\item{externalModules}{file path to directory with external modules}
}
\value{
vector of module classpaths
}
\description{
Get a simple list of all BioModules available to BioLockJ that are also API modules.
}
\examples{
listApiModules()

}
\seealso{
list_modules
}
