% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listModules.R
\name{listModules}
\alias{listModules}
\title{List Modules}
\usage{
listModules(externalModules = NULL)
}
\arguments{
\item{externalModules}{file path to directory with external modules}
}
\value{
vector of module classpaths
}
\description{
Get a simple list of all BioModules available to BioLockJ.
}
\examples{
listModules()

}
\seealso{
listApiModules

moduleInfo
}
