% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMounts.R, R/listUploads.R
\name{listMounts}
\alias{listMounts}
\alias{listUploads}
\title{List Required Files}
\usage{
listMounts(config)

listUploads(config)
}
\arguments{
\item{config}{The BioLockJ config file (or properties file) to scan for required files/dirs.}
}
\value{
listMounts: a list of directories that would need to be mounted in order for
the files listed in the config file to be available to a pipeline running in docker.

listUploads: a list of file and directories that would need to be uploaded in order for
the files listed in the config file to be available to a pipeline running in the cloud.
}
\description{
Create a list of files and/or dirs that a pipeline will need to have access to based on the config file.
}
\section{Functions}{
\itemize{
\item \code{listMounts}: 

\item \code{listUploads}: 
}}

