% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAllProps.R, R/listProps.R
\name{listAllProps}
\alias{listAllProps}
\alias{listProps}
\title{List BioLockJ Properties}
\usage{
listAllProps(externalModules = NULL)

listProps(module = NULL)
}
\arguments{
\item{externalModules}{file path to directory with external modules}

\item{module}{If specified, the property list returned will be the properties used by this module.}
}
\value{
character vector of property names

character vector of property names
}
\description{
Display the properties currently used by the BioLockJ java program. Optionally specify a module.
}
\details{
If no module is specified, the properties returned are the general properties.
If a module is specified, then only the properties used by that module are returned,
which usually has some overlap with the general peroperties.

listAllProps() is equivalent to the combined set of listProps() with no args and
listProps(module) for all available modules.
}
\section{Functions}{
\itemize{
\item \code{listAllProps}: 

\item \code{listProps}: 
}}

\examples{
listProps()

}
