% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleInfo.R
\name{moduleInfo}
\alias{moduleInfo}
\title{Get Module Info}
\usage{
moduleInfo(externalModules = NULL)
}
\arguments{
\item{externalModules}{file path to directory with external modules}
}
\value{
a list of moduleInfo objects,
each has named values for module aspects include a named list of property objects
very similar to the output of propInfo().
}
\description{
Get list of modules currently available to the BioLockJ java program.
}
\examples{
moduleInfo()

}
\seealso{
listModules

listApiModules

propInfo
}
