% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameProperties.R
\name{nameProperties}
\alias{nameProperties}
\title{Property objects are named}
\usage{
nameProperties(propList)
}
\arguments{
\item{propList}{a list of property objects}
}
\value{
the propList, as a named list
}
\description{
Property objects are named
}
