% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseListProp.R
\name{parseListProp}
\alias{parseListProp}
\title{Separate elements of a BioLockJ list property}
\usage{
parseListProp(value)
}
\arguments{
\item{value}{the single-string value a property}
}
\value{
a character vector of the separated elements of the list
}
\description{
Separate elements of a BioLockJ list property
}
\examples{
parseListProp("one, two, three")
#>[1] "one"   "two"   "three"

}
\seealso{
printListProp
}
