% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printListProp.R
\name{printListProp}
\alias{printListProp}
\title{Assemble a BioLockJ list property}
\usage{
printListProp(values)
}
\arguments{
\item{values}{a vecotr of values to combine as one list property}
}
\value{
a single string representing the set of values as a comma-separated list
}
\description{
Assemble a BioLockJ list property
}
\examples{
printListProp(c("one", "two", "three"))
#> [1] "one, two, three"

}
\seealso{
parseListProp
}
