% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeProp.R, R/propInfo.R, R/propType.R,
%   R/propValue.R
\name{describeProp}
\alias{describeProp}
\alias{propInfo}
\alias{propType}
\alias{propValue}
\title{Get Information about BioLockJ Properties}
\usage{
describeProp(property, module = NULL, externalModules = NULL)

propInfo()

propType(property, module = NULL, externalModules = NULL)

propValue(property, config = NULL, module = NULL, externalModules = NULL)
}
\arguments{
\item{property}{the name of a property}

\item{module}{a BioLockJ module}

\item{externalModules}{file path to directory with external modules}

\item{config}{a BioLockJ config file (or properties file) to reference}
}
\value{
describeProp: the description of the property as a character string

propInfo: a list of property objects, where each element has named values for "property" (ie name), "type", and "description" and (when applicable) "default".

propType: the type of the property as a character string

propValue: the value of the property (iether the default or the value as specified in config)
}
\description{
Get Information about BioLockJ Properties
}
\section{Functions}{
\itemize{
\item \code{describeProp}: 

\item \code{propInfo}: 

\item \code{propType}: 

\item \code{propValue}: 
}}

\examples{
describeProp("script.numThreads")
#> "Used to reserve cluster resources and passed to any external application call that accepts a numThreads parameter."

propType("script.numThreads")
#> [1] "integer"

propValue("script.numThreads")
#> "8"

propValue("hardyHarHar") # unrecognized property
#> ""

}
