% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_properties.R, R/read_props_file.R
\name{read_properties}
\alias{read_properties}
\alias{read_props_file}
\title{Read BioLockJ properties}
\usage{
read_properties(existingLines)

read_props_file(filepath)
}
\arguments{
\item{existingLines}{a character vector, usually the result of readLines()}

\item{filepath}{character string, exactly one, giving the path to the file to read.}
}
\value{
read_properties: returns a named character vector.  Names are property names (left of =), values are property values (right of =).

read_props_file: is a wrapper for read_properties that takes a file; returns result of read_properties().
}
\description{
Read properties from a BioLockJ config file or properties file, or a vector of lines representing BioLockJ properties.
}
\section{Functions}{
\itemize{
\item \code{read_properties}: 

\item \code{read_props_file}: 
}}

\examples{
ex1 = system.file("extdata", "example_with_defaults.properties", package = "BioLockR")
read_props_file(ex1)

ex2 = system.file("extdata", "example_no_defaults.properties", package = "BioLockR")
read_properties( readLines(ex2) )
# identical to:
read_props_file(ex2)

vals = c("name=value", "property.name=its value", "property.other=another value")
read_properties( vals )

}
