% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBljJar.R, R/setBljProj.R
\name{setBljJar}
\alias{setBljJar}
\alias{setBljProj}
\title{Set BioLockJ Essential File Paths}
\usage{
setBljJar(path, remember = TRUE, doublecheck = TRUE)

setBljProj(path, remember = TRUE, doublecheck = TRUE)
}
\arguments{
\item{path}{a file path to the BioLockJ jar file (setBljJar()) to the BioLockJ project directory (setBljProj).}

\item{remember}{TRUE/FALSE: Should this value be remembered for future sessions.
If true, the value is appended to the $HOME/.Renviron file, which can be edited
using: usethis::edit_r_environ()}

\item{doublecheck}{(only used if remember==TRUE) should an interactive response be required before saving value to a file.}
}
\value{
TRUE if path is set for current session


}
\description{
setBljJar and setBljProj set variables BLJ_JAR and BLJ_PROJ respectively.
These variables can also be set by adding them to your .Renviron file, which
can be edited using: \code{usethis::edit_r_environ()} .
}
\section{Functions}{
\itemize{
\item \code{setBljJar}: 

\item \code{setBljProj}: 
}}

\examples{
setBljJar("path/to/BioLockJ/dist/BioLockJ.jar")

setBljProj( "path/to/projects/dir" )

}
\seealso{
getBljJar, getBljProj
}
