% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_properties.R
\name{write_properties}
\alias{write_properties}
\title{Write Properties}
\usage{
write_properties(file = NULL, properties = NULL, moduleLines = NULL)
}
\arguments{
\item{file}{file location to right properties to}

\item{properties}{named vector or named list of property values}

\item{moduleLines}{character vector representing a module run order for a BioLockJ configuration}
}
\value{
If file is NULL, returns a character vector of the lines that would make up the config file;
if file is given, the lines are written to the file and the function returns TRUE.
}
\description{
Write properties in the format expected by BioLockJ. This does not incorporate any human-readable nicities.
}
