% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callBioLockJApi.R
\name{callBioLockJApi}
\alias{callBioLockJApi}
\title{Call the helper API of BioLockJ java program}
\usage{
callBioLockJApi(args, external.modules = NULL, useProj = FALSE)
}
\arguments{
\item{args}{args to pass to BioLockJ java program}

\item{external.modules}{file path to directory with external modules}

\item{useProj}{boolean, should the vaue of getBljProj() be passed to the system call as an environment variable.
Unlike the main BioLockJ command, most of the API commands do not use BLJ_PROJ, and can run before that value is set.}
}
\value{
Any and all results that the BioLockJ java program prints to standard out
}
\description{
Call the helper API of BioLockJ java program
}
