% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_BioLockR.R, R/getBljProj.R,
%   R/get_BLJ_JAR.R, R/setBljProj.R, R/set_BLJ_JAR.R, R/setup_BioLockR.R
\docType{package}
\name{config_BioLockR}
\alias{config_BioLockR}
\alias{getBljProj}
\alias{get_BLJ_JAR}
\alias{setBljProj}
\alias{set_BLJ_JAR}
\alias{setup_BioLockR}
\title{Configure BioLockR}
\usage{
getBljProj()

get_BLJ_JAR()

setBljProj(path, remember = TRUE, doublecheck = TRUE)

set_BLJ_JAR(path, remember = TRUE, doublecheck = TRUE)

setup_BioLockR()
}
\arguments{
\item{path}{a file path to the BioLockJ jar file (set_BLJ_JAR()) to the BioLockJ project directory (setBljProj).}

\item{remember}{TRUE/FALSE: Should this value be remembered for future sessions.
If true, the value is appended to the $HOME/.Renviron file, which can be edited
using: usethis::edit_r_environ()}

\item{doublecheck}{(only used if remember==TRUE) should an interactive response be required before saving value to a file.}
}
\value{
"get" methods return a file path.
"set" methods, return boolean; TRUE if the value is set for that property

getBljProj: the path to the BioLockJ project directory

get_BLJ_JAR: The path to the BioLockJ jar file

setBljProj: boolean; TRUE if path is set for current session

set_BLJ_JAR: boolean; TRUE if path is set for current session

setup_BioLockR: no returned value
}
\description{
Set BioLockJ Essential File Paths

This command launches an interactive mode that promps the user for the required settings.
}
\details{
The BioLockR package is way to interact with the BioLockJ java program.
At a minimim, it requires java and the path to the jar file in the BioLockJ folder:
.../BioLockJ/dist/BioLockJ.jar

set_BLJ_JAR and setBljProj set variables BLJ_JAR and BLJ_PROJ respectively.

setup_BioLockR
setup_BioLockR() launches an interactive mode that promps the user for the required settings.
The same setup can be done prgrammatically by using set_BLJ_JAR() and setBljProj() with doublecheck=FALSE.

set_BLJ_JAR() and setBljProj() will prompt about using the settings for
future sessions (unless doublecheck=FALSE).  This interactive session is designed for an initial setup,
and automatically uses remember=TRUE and doublecheck=FALSE when seting BLJ_JAR and BLJ_PROJ.

Unlike set_BLJ_JAR(), this method checks for java and also tests the jar file by attempting to get the BioLockJ version from it.
This check is skipped if java is not accessible.

If you have BioLockJ installed, you have a variable $BLJ, and likley $BLJ_RPOJ.
These may not be detected in an R session, or an Rstudio R session.
If these are detected, they are used; otherwise they must be set manually.

In some cases, an R session launched from the terminal may have access to the
variables, while a session launched in Rstudio does not. In this case,
run setup_BioLockR() from the terminal R session, and subsequent sessions in Rstudio will have the paths.
}
\section{Functions}{
\itemize{
\item \code{getBljProj}: get the BioLockJ projects directory path

\item \code{get_BLJ_JAR}: get the BioLockJ.jar file path

\item \code{setBljProj}: set the BioLockJ projects directory path

\item \code{set_BLJ_JAR}: set the BioLockJ.jar file path

\item \code{setup_BioLockR}: interactively set all configurations.
}}

\examples{
# setBljProj("~/Documents/BioLockJ/pipelines")
# getBljProj()
# >[1] "~/Documents/BioLockJ/pipelines"

# set_BLJ_JAR("/path/to/BioLockJ/dist/BioLockJ.jar")
# get_BLJ_JAR()

setBljProj( "path/to/projects/dir" )

set_BLJ_JAR("path/to/BioLockJ/dist/BioLockJ.jar")

}
\seealso{
set_BLJ_JAR

setBljProj
}
