% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClassPath.R
\name{getClassPath}
\alias{getClassPath}
\title{Get Class Path for BioLockJ call}
\usage{
getClassPath(externalModules = NULL)
}
\arguments{
\item{externalModules}{optional path to directory containing jar files for additional modules}
}
\value{
the class path string to use when calling BioLockJ
}
\description{
Get Class Path for BioLockJ call
}
