% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasReadableValue.R, R/isReadableValue.R
\name{hasReadableValue}
\alias{hasReadableValue}
\alias{isReadableValue}
\title{Is Readable Value}
\usage{
hasReadableValue(value)

isReadableValue(value)
}
\arguments{
\item{value}{usually the value of a property or a user response.}
}
\value{
a single boolean. Return FALSE if value is NULL, NA, empty vector/list or empty string; otherwise return TRUE.

boolean. Return FALSE if value is NULL, NA, empty vector/list or empty string; otherwise return TRUE.
}
\description{
Use this method to prevent failures from NULL, NA, or empty values.
}
\section{Functions}{
\itemize{
\item \code{hasReadableValue}: return a single boolean, recommended for logic flow.

Use this method to prevent failures from NULL, NA, or empty values.

\item \code{isReadableValue}: result may be a vector or matrix reflecting the structure of \emph{value}; recommended for subsetting.
}}

\examples{
hasReadableValue(NA)
#> FALSE
hasReadableValue(c())
#> FALSE
hasReadableValue("")
#> FALSE
hasReadableValue("text")
#> TRUE
# hasReadableValue(c(1,2,3,NA,4))
# #> TRUE
# hasReadableValue(list(a=c(1,2,NA), b=c("app"), c=NA, d=""))
# #> TRUE

isReadableValue(NA)
#> FALSE
isReadableValue(c())
#> FALSE
isReadableValue("")
#> FALSE
isReadableValue("text")
#> TRUE
# isReadableValue(c(1,2,3,NA,4))
# #> TRUE  TRUE  TRUE FALSE  TRUE
# isReadableValue(list(a=c(1,2,NA), b=c("app"), c=NA, d=""))
# #> $a
# #> [1]  TRUE  TRUE FALSE
# #> $b
# #> [1] TRUE
# #> $c
# #> [1] FALSE
# #> $d
# #> [1] FALSE

}
