% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isReadableValue.R
\name{isReadableValue}
\alias{isReadableValue}
\title{Is Readable Value}
\usage{
isReadableValue(value)
}
\arguments{
\item{value}{usually the value of a property or a user response.}
}
\value{
boolean. Return FALSE if value is NULL, NA, empty vector/list or empty string; otherwise return TRUE.
}
\description{
Use this method to prevent failures from NULL, NA, or empty values.
}
\examples{
isReadableValue(NA)
#> FALSE
isReadableValue(c())
#> FALSE
isReadableValue("")
#> FALSE
isReadableValue("text")
#> TRUE

}
