% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_BioLockR.R
\name{setup_BioLockR}
\alias{setup_BioLockR}
\title{Setup required file paths}
\usage{
setup_BioLockR()
}
\value{
no returned value
}
\description{
This command launches an interactive mode that promps the user for the required settings.
}
\details{
The BioLockR package is way to interact with the BioLockJ java program.
At a minimim, it requires java and the path to the jar file in the BioLockJ folder:
.../BioLockJ/dist/BioLockJ.jar

This command launches an interactive mode that promps the user for the required settings.
The same setup can be done prgrammatically by using setBljJar() and setBljProj() with doublecheck=FALSE.

setBljJar() and setBljProj() will prompt about using the settings for
future sessions (unless doublecheck=FALSE).  This interactive session is designed for an initial setup,
and automatically uses remember=TRUE and doublecheck=FALSE when seting BLJ_JAR and BLJ_PROJ.

Unlike setBljJar(), this method checks for java and also tests the jar file by attempting to get the BioLockJ version from it.
This check is skipped if java is not accessible.
}
\seealso{
setBljJar

setBljProj
}
